requirements.txt
streamlit
slack-sdk
python-dotenv
black
pylint
